move custom hook to helpers.ts
