+++ src/contexts/helpers.ts
@@ -1,0 +1,6 @@
+import React from "react";
+import { SettingsContext } from "./SettingsContext";
+
+export function useSettingsContext() {
+    return React.useContext(SettingsContext);
+}
